Extract get_base_url and get_full_url utils outside of api.v2
