+++ wagtail/api/v2/signal_handlers.py
@@ -1,13 +1,12 @@
 from django.db.models.signals import post_delete, post_save
 from django.urls import reverse
 
+from wagtail.api.utils import get_base_url
 from wagtail.contrib.frontend_cache.utils import purge_url_from_cache
 from wagtail.documents import get_document_model
 from wagtail.images import get_image_model
 from wagtail.models import get_page_models
 from wagtail.signals import page_published, page_unpublished
-
-from .utils import get_base_url
 
 
 def purge_page_from_cache(instance, **kwargs):
