+++ wagtail/api/v2/tests/tests.py
@@ -1,11 +1,11 @@
 from django.test import RequestFactory, TestCase, override_settings
 from django.utils.encoding import force_bytes
 
+from wagtail.api.utils import get_base_url
 from wagtail.models import Site
 
 from ..utils import (
     FieldsParameterParseError,
-    get_base_url,
     parse_boolean,
     parse_fields_parameter,
 )
