+++ wagtail/api/v2/utils.py
@@ -1,39 +1,13 @@
-from urllib.parse import urlsplit
-
 import swapper
-from django.conf import settings
-from django.utils.encoding import force_str
 
+from wagtail.api.utils import get_full_url
 from wagtail.coreutils import resolve_model_string
-from wagtail.models import Site
 
 Page = swapper.load_model("wagtailcore", "Page")
 
 
 class BadRequestError(Exception):
     pass
-
-
-def get_base_url(request=None):
-    base_url = getattr(settings, "WAGTAILAPI_BASE_URL", None)
-
-    if base_url is None and request:
-        site = Site.find_for_request(request)
-        if site:
-            base_url = site.root_url
-
-    if base_url:
-        # We only want the scheme and netloc
-        base_url_parsed = urlsplit(force_str(base_url))
-
-        return base_url_parsed.scheme + "://" + base_url_parsed.netloc
-
-
-def get_full_url(request, path):
-    if path.startswith(("http://", "https://")):
-        return path
-    base_url = get_base_url(request) or ""
-    return base_url + path
 
 
 def get_object_detail_url(router, request, model, pk):
